refactor(BUX-497): move GenerateServiceURL to capabilities as private method
