+++ server/config_test.go
@@ -303,50 +303,6 @@ func TestConfiguration_EnrichCapabilities(t *testing.T) {
 	})
 }
 
-// TestGenerateServiceURL will test the method GenerateServiceURL()
-func TestGenerateServiceURL(t *testing.T) {
-	t.Parallel()
-
-	t.Run("valid values", func(t *testing.T) {
-		u, err := GenerateServiceURL("https://", "test.com", "v1", "bsvalias")
-		assert.NoError(t, err)
-		assert.Equal(t, "https://test.com/v1/bsvalias", u)
-	})
-
-	t.Run("all invalid values", func(t *testing.T) {
-		_, err := GenerateServiceURL("", "", "", "")
-		assert.Error(t, err)
-	})
-
-	t.Run("missing prefix", func(t *testing.T) {
-		_, err := GenerateServiceURL("", "test.com", "v1", "")
-		assert.Error(t, err)
-	})
-
-	t.Run("missing domain", func(t *testing.T) {
-		_, err := GenerateServiceURL("https://", "", "v1", "")
-		assert.Error(t, err)
-	})
-
-	t.Run("no api version", func(t *testing.T) {
-		u, err := GenerateServiceURL("https://", "test", "", "bsvalias")
-		assert.NoError(t, err)
-		assert.Equal(t, "https://test/bsvalias", u)
-	})
-
-	t.Run("no service name", func(t *testing.T) {
-		u, err := GenerateServiceURL("https://", "test", "v1", "")
-		assert.NoError(t, err)
-		assert.Equal(t, "https://test/v1", u)
-	})
-
-	t.Run("service with explicit port", func(t *testing.T) {
-		u, err := GenerateServiceURL("https://", "test:1234", "v1", "bsvalias")
-		assert.NoError(t, err)
-		assert.Equal(t, "https://test:1234/v1/bsvalias", u)
-	})
-}
-
 // TestNewConfig will test the method NewConfig()
 func TestNewConfig(t *testing.T) {
 	t.Parallel()
